purge cdn after SDK connection changes
