+++ packages/back-end/src/services/features.ts
@@ -1,7 +1,6 @@
 import { webcrypto as crypto } from "node:crypto";
 import { createHash } from "crypto";
 import uniqid from "uniqid";
-import fetch from "node-fetch";
 import isEqual from "lodash/isEqual";
 import omit from "lodash/omit";
 import { orgHasPremiumFeature } from "enterprise";
@@ -34,7 +33,7 @@ import { queueProxyUpdate } from "../jobs/proxyUpdate";
 import { ApiFeature, ApiFeatureEnvironment } from "../../types/openapi";
 import { ExperimentInterface, ExperimentPhase } from "../../types/experiment";
 import { VisualChangesetInterface } from "../../types/visual-changeset";
-import { FASTLY_API_TOKEN, FASTLY_SERVICE_ID } from "../util/secrets";
+import { purgeCDNCache } from "../util/fastly.util";
 import { getEnvironments, getOrganizationById } from "./organizations";
 
 export type AttributeMap = Map<string, string>;
@@ -264,33 +263,6 @@ export function getSurrogateKey(
 
   // Protect against environments or projects having unusual characters
   return key.replace(/[^a-zA-Z0-9_-]/g, "");
-}
-
-export async function purgeCDNCache(
-  orgId: string,
-  payloadKeys: SDKPayloadKey[]
-): Promise<void> {
-  // Only purge when Fastly is used as the CDN (e.g. GrowthBook Cloud)
-  if (!FASTLY_SERVICE_ID || !FASTLY_API_TOKEN) return;
-
-  // Only purge the specific payloads that are affected
-  const surrogateKeys = payloadKeys.map((k) =>
-    getSurrogateKey(orgId, k.project, k.environment)
-  );
-  if (!surrogateKeys.length) return;
-
-  try {
-    await fetch(`https://api.fastly.com/service/${FASTLY_SERVICE_ID}/purge`, {
-      method: "POST",
-      headers: {
-        "Fastly-Key": FASTLY_API_TOKEN,
-        "surrogate-key": surrogateKeys.join(" "),
-        Accept: "application/json",
-      },
-    });
-  } catch (e) {
-    logger.error("Failed to purge cache for " + orgId);
-  }
 }
 
 export type FeatureDefinitionsResponseArgs = {
